players query
players: takes leagueCode as a parameter and returns the players that belong to all teams participating in the given league. If the given leagueCode is not present in the DB, it should respond with an error message. Add an optional input to the query to filter players also by team name. Important: If there is no players’ data available (see Note above), then make the same implementation for this using coaches instead.